Use explicit device_release primitive in the frontend (#1720)
Currently, we only instantiate quantum dealloc primitives in the JAXPR,
which are automatically lowered to a `quantum.dealloc` and
`quantum.device_release` operation in MLIR. This is a blocker for
dynamic allocation since we don't want to have to release the device
when deallocating a register, as that would destroy any other qubits
that are still active as well.

The PR is a small fix to split the deallocation primitive into explicit
deallocation and device release primitives. In Catalyst they are still
called together.

fixes #1701