Update clang-format and clang-tidy dependancy (#1721)
**Context:**
As part of updating dependancy versions for the ecosystem, we want to
update formatting tools like `clang-tidy` and `clan-format`.

**Description of the Change:**
Bump `clang-tidy` and `clang-format` to v20.

Note that I changed the Formatting CI job to install clang-format from
the requirements this makes the job consistent with local development
environment and allow us to use the latest version that are not yet
available on the ubuntu repositories.

**Benefits:**
Using the latest formatting tools

**Possible Drawbacks:**

**Related GitHub Issues:**
[sc-91016]