Migrate `catalyst` dialect to new one-shot bufferization (#1708)
**Context:**
This work is based on #1027 .

As part of the mlir update, the bufferization of the custom catalyst
dialects need to migrate to the new one-shot bufferization interface, as
opposed to the old pattern-rewrite style bufferization passes.
See more context in https://github.com/PennyLaneAI/catalyst/pull/1027.

The `Quantum` dialect was migrated in #1686 .

**Description of the Change:**
MIgrate `Catalyst` dialect to one-shot bufferization.

**Benefits:**
Align with mlir practices; one step closer to updating mlir.

[sc-71487]

---------

Co-authored-by: Tzung-Han Juang <tzunghan.juang@gmail.com>