Improve warning message for clang-format version mismatch with CI (#1745)
There's a warning message in the formatting script that warns a user
running `make format` when their local version of `clang-format` is
older than the version used in CI. Currently, this warning message is
displayed as

```
Using clang-format version 14.                     As this is lower than the version used for the CI,                     the CI may fail even after formatting.
```

The multi-line string in the source code is displayed with extra white
space on a single line in the terminal. This PR removes the extra white
space, and also adds the `clang-format` version used in the CI to the
message. For example, the message would now be:

```
Warning: using clang-format version 14, while the CI uses version 20. Since the local version is lower than the version used for the CI, the CI may fail even after formatting.
```