📦️ Export Headers for the MLIRQuantum Dialect in the Wheel (#1733)
**Context:**

When building an MLIR Plugin to be used in combination with Catalyst and
Pennylane, during the build process the header files for the Quantum
dialect are required. Those are only generated at the build time of the
`mlir` target.

**Description of the Change:**

The major change of the PR is that when building the wheel, the header
corresponding to the MLIR dialect are copied to a destination such that
it can be included in the wheel. Additionally, the `catalyst` python
module receives a command line interface to retrieve the location of the
header files by running `python -m catalyst --include_dir`.

Secondly, this PR removes `nlohmann::json` from the exported targets.
This is actually not necessary anymore, as only the header files are
used for building the standalone plugin. Previously, instead of
including the headers, we linked to the target MLIRQuantum. Doing that
in a project like MQT Core which already has `nlohmann::json` as a
dependency, this leads to an error because of duplicate symbols. Hence,
I think this change is still valuable even though not mandatory for the
main purpose of this PR.

**Benefits:**

MLIR plugins can be built by just installing `pennylane-catalyst` into
the Python environment without cloning and building the entire catalyst.

**Possible Drawbacks:**

This can only be tested by building the wheel, installing the wheel in
the environment, and then trying to build the MLIR plugin using the
header from the wheel. I do not know, how this could be integrated into
the current testing framework.

**Related GitHub Issues:**

none

---------

Co-authored-by: Erick Ochoa Lopez <erick.ochoalopez@xanadu.ai>