Add resource tracking to NullQubit (#1619)
**Context:**
For benchmarking, we need a method of precise resource tracking. This
method has to be compatible with code run directly from PennyLane and
should be able to track precise resources after various transformation
passes have already occurred.

See also: https://github.com/PennyLaneAI/pennylane-benchmarks/pull/76
and https://github.com/PennyLaneAI/pennylane/pull/7226

**Description of the Change:**
Adds 2 new class attributes to `NullQubit`, as well as an optional
argument to its constructor. When this option is toggled on, `NullQubit`
will track precise gate counts while "executing" a circuit and save them
to a file when all qubits are released.

**Benefits:**
It is now possible to track precise gate counts (with or without gate
decomposition or other transformation passes) for measuring circuit
complexity.

**Possible Drawbacks:**
This implementation currently write the resources JSON to a file on disk
as opposed to neatly returning an object back through to the Python
caller in PennyLane. In the future, a more robust solution may be
desired.

**Related GitHub Issues:**
[sc-91366]

---------

Co-authored-by: Ali Asadi <10773383+maliasadi@users.noreply.github.com>
Co-authored-by: David Ittah <dime10@users.noreply.github.com>