Migrate `gradient` dialect to new one-shot bufferization (#1740)
**Context:**
This work is based on #1027 .

As part of the mlir update, the bufferization of the custom catalyst
dialects need to migrate to the new one-shot bufferization interface, as
opposed to the old pattern-rewrite style bufferization passes.
See more context in #1027.

The `Quantum` dialect was migrated in #1686 .
The `Catalyst` dialect was migrated in #1708 .

Note that #1139 refactors the gradient dialect's bufferization into
preprocess, bufferization, and postprocess.
Only the middle bufferization stage is supposed to be replaced by
one-shot bufferization.

**Description of the Change:**
Migrate `Gradient` dialect to one-shot bufferization.

**Benefits:**
Align with mlir practices; one step closer to updating mlir.

[sc-71487]

---------

Co-authored-by: Tzung-Han Juang <tzunghan.juang@gmail.com>