Fix adjoint gradient.
New pattern applicator (`applyPatternsGreedily`) is more aggressive in dead code elimination,
and is eliminating dead `Value`s in the adjoint gradient method.
The `nodealloc` function used to only return the qreg, not the expval result.
This causes the expval op to be eliminated since it has no users.
This further causes wrong gradient results, since the entire program, all ops included (regardless
of dead or not), impacts the gradient.

To avoid this, we return the expval result as well.