As part of the llvm update, the old --buffer-deallocation pass is removed. Intended replacement is --buffer-deallocation-pipeline. [mlir][bufferization] Remove buffer-deallocation pass llvm/llvm-project#126366, https://discourse.llvm.org/t/psa-bufferization-new-buffer-deallocation-pipeline/73375. However, we encountered a few issues. See detailed discord discussion at https://discord.com/channels/636084430946959380/642426447167881246/1376919538301403276
The heart of the new deallocation pass, --ownership-based-buffer-deallocation, blanket-ly fails for ops with unknown memory effects: https://github.com/llvm/llvm-project/blob/da4958ae2b384c2a027cf20c67b7e211d39fcbfe/mlir/lib/Dialect/Bufferization/Transforms/OwnershipBasedBufferDeallocation.cpp#L523
To solve this issue, the suggestion was to add memory effects to custom operations.
The migration guide also suggests that bufferizable ops no longer implement the bufferizesToAllocation method, so we remove them.

This was supposed to be done alongside the llvm update in https://github.com/PennyLaneAI/catalyst/pull/1752;
However, soon it became clear that this migration to the new buffer deallocation is very complicated, and should be its own story.

The llvm update in #1752 thus did not finish this migration.
This PR records the work that was already done on this back in #1752.