Add more OQD tests (#1763)
**Context:**
OQD openAPL tests are limited to a single RX gate

**Description of the Change:**
- [x] add proper cleanup fixture

Expland the test to cover:
- [x] Target gate set {RX, RY, MS(not added yet)}
- [x] Gates for which there is a decomposition to the target gate set
{CNOT, RZ, PhaseShift, Hadamard, S, T, PauliZ}
- [x] More complex circuits (RX_RY, 2-qubit QFT)

Modified OQD toml file to ensure more complicated circuits like the QFT
are only decomposed to the {CNOT, RZ, PhaseShift, Hadamard, S, T,
PauliZ} which are decomposable to the target gate set.

To avoid a situation where different tests compete over the same openAPL
output file, we pass the name of the opanAPL file as kwarg to the OQD
device which is used in runtime to make the json file.

**Benefits:**

**Possible Drawbacks:**
Huge number of lines need to be added in order to compare the whole
openAPL json files. Alternatively, we only check for more high level
infos (e.g. number of Pulses, specific parameters, etc.) instead of
adding the huge number of lines for openAPL programs

**Related GitHub Issues:**
[sc-88299]

---------

Co-authored-by: Paul <79805239+paul0403@users.noreply.github.com>