Implement automatic qubit management (#1788)
**Context:**
We wish to support automatic qubit management, i.e. when the user does
not specify the number of wires in device initialization
```python
@qjit
def workflow():
    dev = qml.device("lightning.qubit") # no wires here!
    @qml.qnode(dev)
    def circuit():
        qml.PauliX(wires=0)
        return qml.probs(wires=[2,3])  # non qjit is [1,0,0,0]
    return circuit()
```

**Description of the Change:**
A new unit attribute, `auto_qubit_management`, is added to the device
init op. If no wires are specified during device init in user python, we
trace to device init op with this attribute turned on.

In runtime, when a new user wire index is encountered in
`__catalyst__rt__array_get_element_ptr_1d`, if the device is in
automatic management mode (as indicated by the `RTDevice` constructor,
set from the new attribute we added), we allocate new device qubits so
that this new user wire index is an available qubit.

**Benefits:**
The automatic qubit management feature is now supported.

**Possible Drawbacks:**
1. Catalyst and Pennylane interprets wire labels differently. See #1800 
2. Potential out-of-memory error might be encountered in runtime.

[sc-86500]