[Frontend] Split `QFuncPlxprInterpreter` into `SubroutineInterpreter` (#1787)
**Context:** A quantum subroutine and the quantum kernel's entry point
have almost everything in common. The only differences are:

* the kernel's entry point will define the quantum register as one of
the first few instructions in its body, while the quantum subroutine the
quantum register will be passed in as a parameter.
* the kernel's entry point will deallocate the quantum register as one
if its last few instruction in its body, while the quantum subroutine
will return the most recently updated value of the quantum register.
* the kernel's entry point will deallocate the device as one of its last
few instruction in its body.

**Description of the Change:** Move all the semantics of the quantum
instructions except the setup and teardown method's for
QFuncPlxprInterpreter into `SubroutineInterpreter`.

**Benefits:** Avoid code duplication once support for subroutines are
added in the frontend.