Add flaky pytest infrastructure and weekend runners  (#1799)
**Context:**
We wish to reduce flaky tests in CI.
There is a python package `flaky` (which we already have in our
`requirements.txt`) that reruns failed tests to hunt for flakiness.
Let's actually use it.

**Description of the Change:**
- Add `flaky` infrastructure. A pytest run that hunts for flaky tests
can be launched by `make pytest ENABLE_FLAKY=ON`
- 4 tests that change some global state cannot be run more than once. We fix their issues.
- Add a script to launch flaky hunting runs on weekends. 

**Benefits:**
If any new tests are flaky, this weekend runner will catch them!

[sc-87028]

---------

Co-authored-by: David Ittah <dime10@users.noreply.github.com>