Frontend: Refactoring of PLxPR to JAXPR with quantum primitives pipeline (#1805)
**Context:** It was difficult for me to understand when to create
BranchPlxprInterpreter or PredicatePlxprInterpreter or another kind of
interpreter. It made it difficult to reason about the capabilities of
each different interpreter. Turns out, the only difference was the
calling convention, which can easily be coded into the function that
will be traced.

**Description of the Change:** Remove different PlxprInterpreters and
substitute it with a single one that transforms Plxpr into JAXPR with
quantum primitives.

**Benefits:** Subjectively easier to understand. (No need for that
`__getattr__` `__setattr__` hack).

---------

Co-authored-by: paul0403 <paulhaochen.wang@gmail.com>