use `dataclass.replace` to update `ExecutionConfig` or `MCMConfig` (#1814)
**Context:**

We are freezing the dataclasses `ExecutionConfig` or `MCMConfig` in
Pennylane (https://github.com/PennyLaneAI/pennylane/pull/7697) so they
are no longer mutable. Turns out there are external Catalyst tests that
failed since there are some files in this repo that are mutating the
configuration.

**Description of the Change:**

Use `dataclass.replace` to update various properties of the dataclass so
we don't mutate.

**Benefits:** Not in place mutating the execution configuration.

**Possible Drawbacks:** None identified.

[sc-93592]