hoist the pulse duration calculation arithmetic out of the parallel protocol (#1818)
**Context:**
Currently, the `quantum-to-ion` pass lowers the gates to patterns of
ParallelProtocol + pulses. The semantic of ParallelProtocol is that all
the instructions inside it are executed in parallel. However, we are
including some arithmetic that calculates the duration of pulses within
the Parallel protocol, therefore creating data dependency between the
parallel instructions.
Moreover, the name of the pass `quantum-to-ion` implies a full
conversion from quantum dialect to ion dialect which is not the case. We
only convert quantum gates and all the other quantum operations remain
intact after the pass is applied.

**Description of the Change:**
- hoisted the pulse duration calculation arithmetic out of the parallel
protocol
- changed the name of the pass to `gates-to-pulses`

**Benefits:**
This prepares the ion dialect for future development.

**Possible Drawbacks:**

**Related GitHub Issues:**

---------

Co-authored-by: Paul <79805239+paul0403@users.noreply.github.com>