Support Static and Dynamic Variables in PLxPR Programs with QJIT (#1810)
**Context:**
Control over static variables in compilation is a general necessity. At
the moment, with `qml.capture.enable()`, additionally specifying a
static variable in `qjit` does not work. With this epic, we want to add
support for static variables in the plxpr-qjit execution pipeline.

**Description of the Change:**
- Updated the `trace_from_pennylane` function signature to include
`dynamic_argnums`.
- Modified the function implementation to correctly handle dynamic
arguments when calling `from_plxpr`.

 
[[sc-93318]]

---------

Co-authored-by: paul0403 <paulhaochen.wang@gmail.com>