Quantum Subroutine (#1806)
**Context:** Machine generated stablehlo is known to generate functions
in the 10k LOC. MLIR compilation strategy revolves around iterating
through instructions. Since MLIR operations are a linked list, iterating
through this list will trash the cache. A potential solution we found a
long time ago was to reduce the number of instructions. This can be
achieved through the user denoting the use of control flow but there are
still many places where a function would make sense.

**Description of the Change:** This PR introduces the notion of quantum
subroutines. In JAXPR, quantum subroutines are analogous to `pjit_p` but
just with a different primitive name. This will allow the generation of
`func.func` ops during their lowering. This feature is currently only
available through PLxPR. It works as follows:

1. generate a wrapper around functions which will call `jax.jit` on a
given function.
```python

@subroutine
def foo(): ...
```

2. PLxPR does not have quantum registers at all, so during the
conversion of PLxPR to JAXPR with quantum primitives we will insert a
register as a parameter and inline the contents of the subroutine.

```python
def flat_foo(*qreg_args):
   qreg, *args = *qreg_args
   retval = interpreter.evaluate(qreg, args)
   return qreg, *retvals
```

3. Once this gets converted to JAXPR with quantum primitives, then it is
up to Catalyst to lower it to machine code.

**Benefits:** Faster compilation if used, but may limit the runtime
speed in case inlining could optimize the function. I think this is a
good trade off since we know compilation is a bottle neck for machine
generated code.

**Possible Drawbacks:** Only usable from PLxPR.

**Related GitHub Issues:**

- [ ] A decorator to get the MLIR of only quantum subroutines would be
useful for program capture / transformations.
- [ ] Pipeline from PLxPR to Catalyst JAXPR is missing gradient and
callback support, so that is also missing from subroutines
- [ ] Subroutines + Gradients is untested and it is not expected to
work.

---------

Co-authored-by: Paul <79805239+paul0403@users.noreply.github.com>
Co-authored-by: David Ittah <dime10@users.noreply.github.com>