Add a more sophisticated test for from_plxpr qreg manager (#1821)
**Context:**
There are some "requests" for making the UI of the from_plxpr qreg
manager clearer. I added a end-to-end test to illustrate.

**Description of the Change:**
Add an end-to-end test for from_plxpr qreg manager. 
The test is a qubit op, then a qreg op, then a qubit op.

**Benefits:**
More test coverage. 
Illustration of internal object's use. 

**Possible Drawbacks:**
Ideally I want to export something like a `_bind_qubit/qreg_primitive()`
function to automatically update the manager with new qubit and qreg
values. However this was a bit difficult since it is not unified across
primitives which argument in `*args` is the qubits.