Improve keep_intermediate to enable keeping IR after each pass (#1791)
**Context:**
`keep_intermediate` is a boolean flag that if enabled, keeps the
intermediate IR files after each pipeline. However, it is quite common
that a developer needs the IR after a specific pass, which is not
accessible in the current format via `keep_intermediate`.

**Description of the Change:**
Added a third level to the `keep_intermediate` that enables accessing
the IR in a more fine grained format i.e. after each pass.
Note that this is not a breaking change, meaning that the current
behaviour of `keep_intermediate` is kept intact.
With this PR, the user can provide three different levels of printing:
- `False` or `0` or `"none"` (default): No intermediate files are kept.
- `True` or `1` or `"basic"`: Standard intermediate files are kept (IR
after each pipeline).
- `2` or `"debug"`: Standard intermediate files are kept, and IR is
saved after each pass.


**Benefits:**
easier development and debugging

**Possible Drawbacks:**

**Related GitHub Issues:**