[Hack week] Profile C++ code base (RT + Plugins) (#1834)
### Before submitting

Please complete the following checklist when submitting a PR:

- [ ] All new functions and code must be clearly commented and
documented.

- [ ] Ensure that code is properly formatted by running `make format`.
The latest version of black and `clang-format-20` are used in CI/CD to
check formatting.

- [ ] All new features must include a unit test.
Integration and frontend tests should be added to
[`frontend/test`](../frontend/test/),
Quantum dialect and MLIR tests should be added to
[`mlir/test`](../mlir/test/), and
Runtime tests should be added to [`runtime/tests`](../runtime/tests/).

When all the above are checked, delete everything above the dashed
line and fill in the pull request template.


------------------------------------------------------------------------------------------------------------

**Context:**

**Description of the Change:**

**Benefits:**

**Possible Drawbacks:**

**Related GitHub Issues:**