Disallow nested dictionaries the device kwarg string in runtime (#1843)
**Context:**
When there are nested dictionaries in the device kwarg string, the
parsing in runtime silently gives wrong results. #1505

**Description of the Change:**
We explicitly disallow this case.

**Benefits:**
No silent failures.

**Possible Drawbacks:**
Ideally we can parse nested dictionaries with a full AST there. 
But this is too much work for too little gain now, as none of the
production-level devices (aka lightning and openqasm) have
nested-dictionary kwargs.

The only device that has a full "json kwarg" during initialization is
the (still experimental) OQD device (for the ion/phonon specs). However,
that json is parsed through the nlohmann json package. We use json
package there because the json format is very rigid in the OQD device
(since they are fixed from the openapl specs). For a more generic kwarg
string with arbitrary levels of dictionary nested-ness, it is unclear
how to proceed. How to initialize the cpp map container for the parsed
results? How many level of nested `<std::map<string, std::map<string,
...` does there need to be? These are unclear, and would require
development time out of proportions to the gains.

So explicitly failing is good enough imo.

**Related GitHub Issues:** closes #1505 
[sc-84216]