Better checking for runtime device kwargs string not having nested dictionaries (#1846)
**Context:**
#1843 bans nested dictionaries in device kwargs.
I assumed that there was only the global dictionary by asserting the
starting and ending char of the string are `{}`,
but this [breaks some tests in lightning
repo](https://github.com/PennyLaneAI/pennylane-lightning/actions/runs/16075489083/job/45369618669?pr=1192#step:6:304).

This assumption is true for devices going through catalyst frontend, but
for standalone tests this is not true.

**Description of the Change:**
Do an actual nested curly braces check.

**Benefits:**
Tests pass.

**Possible Drawbacks:**
The failing tests in lightning are not actually lightning tests. They
are catalyst runtime tests piggy-backing off the lightning device.
This is because null.qubit did not exist back when we moved the
lightning device's catalyst interface into the lightning repo, so we
moved all the tests.
We should move the catalyst runtime tests back to the catalyst repo in
the future.