Build and upload Rc builds nightly to TestPyPi (#1625)
**Context:**

We currently have an
[action](https://github.com/PennyLaneAI/catalyst/blob/main/.github/workflows/build-nightly-release.yaml)
which builds and uploads catalyst to TestPyPi. We would like to do the
same for RC branches during the the feature freeze week for easier
testing of RC branch.

**Description of the Change:**

Created a similar action that builds RC branch every night for the
duration of the release week. Note that the dates and the RC branch tag
are hardcoded and should be modified before every feature freeze.

**Benefits:**

**Possible Drawbacks:**
Note that the version is bumped by a commit nightly and need to be
changed back to the release version by the release.

**Related GitHub Issues:**

---------

Co-authored-by: erick-xanadu <110487834+erick-xanadu@users.noreply.github.com>
Co-authored-by: Raul Torres <138264735+rauletorresc@users.noreply.github.com>
Co-authored-by: David Ittah <dime10@users.noreply.github.com>