Cherry pick nightly rc wheel upload workflow to rc branch  (#1873)
**Context:**
The original PRs for the nightly rc wheel upload was merged to main but
turns out that we need the `.github/workflows/set_rc_version.py` file to
be in the rc branch since the workflow checks out the rc branch before
reading that file.

**Description of the Change:**
Cherry picked the two PRs that was just merged to main to exist in the
rc branch as well.
Note that all the files are within the `.github/workflows` folder, so it
has no impact on the release itself.

---------

Co-authored-by: erick-xanadu <110487834+erick-xanadu@users.noreply.github.com>
Co-authored-by: Raul Torres <138264735+rauletorresc@users.noreply.github.com>
Co-authored-by: David Ittah <dime10@users.noreply.github.com>