Add analysis for static for loops in ppm_specs (#1863)
**Context:**
Currently, the `ppm_specs` pass increments the PPR count by one whenever
it sees a PPR op.
This causes for loops to be incorrectly counted, since the loop body is
not unrolled.

In other words, both snippets below give the same statistics:
```python
        for i in range(15):
            qml.Hadamard(i)

vs

        for i in range(15000000):
            qml.Hadamard(i)
```

**Description of the Change:**
- Add a utility function `countStaicForloopIterations(Operation *)` that
computes the number of times an operation in a static for loop will be
executed. Effectively this is how many ops there will be in the IR if it
was fully unrolled.
- Update `ppm_specs` pass to count using static for loop info whenever
possible.

**Benefits:**
Better (i.e. "more correct") PPR gate counts.

**Possible Drawbacks:**
We can only do this during compile time for static for loops. 
For while loops, conditionals and for loops with dynamic boundary
`Value`s, this cannot be done.
For now, we just bail out and add one gate. This is ok since now
`ppm_specs` only works with static IR anyway.
But maybe in the future we can print out better specs, e.g. something
like `3N+7`

[sc-94822]