Replace the VLA use with standard C++ code (#1887)
**Description of the Change:**

Previously, we enable VLA in tests to pass the test suite
`runtime/tests/Test_NullQubit.cpp` #1875 But seems it's not the standard
C++, and after discussed, the use of vector here wouldn't affect the
`__catalyst_*` functions. So just wrap it with a vector and pass the
internal data to `MemRefT_double_2d`.