Update jax to 0.6.2 (#1897)
**Context:**
Updating jax from 0.6.0 to 0.6.2. 

**Description of the Change:**
1. A few jaxpr internals started requiring a `source_info` argument.
https://github.com/jax-ml/jax/pull/28103
At any point in the program, the current source info at that point can
be retrieved by `jax._src.source_info.current()`.
   The impacted functions are 
   - `DynamicJaxprTrace.to_jaxpr_tracer, new_arg`
- Primitives with custom staging rules
(https://github.com/jax-ml/jax/pull/29347). For us, this is `sample_p,
counts_p, probs_p, state_p`. For core pennylane this is updated in
https://github.com/PennyLaneAI/pennylane/pull/7887
2. Literals are now inlined into jaxpr equations
https://github.com/jax-ml/jax/pull/28233. This has the following
impacts:
- Firstly, when printing jaxpr, the types of literals are now printed.
This required a few lit-ish tests to update.
- When topoligically sorting equations, the invars to equations are not
necessaily all `Var`s, but can now be `Literal`s as well. In the sorting
algorithm we tap into the internals of `Var`, but this would break if an
invar is a `Literal`. This is actually very easy to fix, as literals do
not effect topological ordering and do not need to be dealt with, so we
just skip over them during sorting
3. In the same above jax commit, during tracing they now drop unused
variables. Concretely, right before exiting `DynamicJaxprTrace.to_jaxpr
(and to_jaxpr2)`, they now call `_drop_unused_var()` on the traced
jaxpr. This is usually fine, but in our control flow's tracing, we trace
the classical part of the inner region first before tracing the quantum
tape (see
[`HybridOp`](https://github.com/PennyLaneAI/catalyst/blob/main/frontend/catalyst/jax_tracer.py#L416)).
This means the classical variables used by the quantum tape, which is
"unused" at the end of the classical tracing, will be dropped, and
breaks quantum tracing. So we patch `_drop_unused_var()` to do nothing
when tracing the classical part of control flow ops.
4. `PytreeDef.make_from_node_data_and_children()` has been removed (with
no replacement) https://github.com/jax-ml/jax/pull/29491. I do this
manually now.


[sc-95175]