Add Layer and Yield operations to QEC dialect (#1917)
**Context:**
Introduced `LayerOp` and `YieldOp` to the QEC dialect, enabling the
representation of a group of mutually commuting operations or operations
that act on different qubits. The `LayerOp` requires a single block that
terminates with a `qec.yield` operation.

**Benefits:**
This addition enhances the expressiveness of the QEC dialect, enabling
circuit depth analysis and possibly leading to layer-based backend
execution.

**Example:**
```mlir
    func.func @layer(%arg0 : !quantum.bit, %arg1 : i1) -> i1{
        %m, %0 = qec.layer(%q0 = %arg0, %c = %arg1) : !quantum.bit, i1 {
            %res, %q_1 = qec.ppm ["Z"](4) %q0 cond(%c): !quantum.bit
            qec.yield %res, %q_1 : i1, !quantum.bit
        }
        func.return %m : i1
    }
```

[[sc-95171]]

---------

Co-authored-by: Joey Carter <joseph.carter@xanadu.ai>