Improve patching location in wheel scripts
Since patches modify the source code, they should be associated to the
steps generating that source code, in particular when storing the
resulting patched code in the GitHub cache.

Previously, the patching was done right before the build step, which can
lead to conflicts if the code is already patched.