Support multi-wires Identity op (#1907)
**Context:**

Support multi-wires qubits for the Identity op. 

**Description of the Change:**

Currently, run with `qml.Identity(wires=[0, 1])` would cause
segmentation fault since the runtime function
`__catalyst__qis__Identity` didn't support the variadic inputs. This
patch allows it to handle multi-wires qubits by accepting the variadic
inputs for this runtime function.

And what this patch does includes:
1. Place the input qubits of `__catalyst__qis__Identity` at the end of
the params to support variadic inputs
2. Reorder the Identity op arguments before converting to the LLVM IR.

The changes made in this patch are relatively trivial. Instead of
isolating the `identity` op, similar to how `multiRZ` is handled, it
remains a custom op. This is because the custom op already supports
variadic input qubits.

While this patch is independent of [Remove assert for Identity gate with
>1 wires in Lightning Qubit
#1212](https://github.com/PennyLaneAI/pennylane-lightning/pull/1212), it
still requires corresponding changes/support on the lightning side to
prevent runtime errors.

**Benefits:**

**Possible Drawbacks:**

Since we did not isolate this op, we have kept it as a custom op. This
approach has its pros and cons. The advantage is that it is a relatively
simple solution that does not require extensive changes to the existing
code. The disadvantage is that the `ConversionPatterns.cpp` (to LLVM)
needs to record the names of these ops to maintain consistency with the
runtime API.

**Related GitHub Issues:** Fixes #1175 

[sc-74894]