Improve error message when using subroutines with quantum operations outside qnodes. (#1932)
**Context:** At the moment a subroutine can be used inside or outside a
qnode. If it is inside, then may contain quantum operations. If it is
outside, then it must not contain quantum operations. When it is outside
and it does contain quantum operations, an exception happens, but the
exception message may not be intuitive for end users.

**Description of the Change:** Improve error message when subroutines
used quantum operations outside of the quantum context.

**Benefits:** Better error messages.

**Possible Drawbacks:** None.

**Related GitHub Issues:**

[sc-96054]