Add catalyst.ftqc.mbqc_pipeline for experimental MBQC workloads (#1942)
**Context:** When we initially added the MBQC dialect, we had to
manually specify the full compilation pipeline add drop in the
MBQC-to-LLVM dialect-conversion pass, `"convert-mbqc-to-llvm"`, to
convert the ops in the `mbqc` dialect to the LLVM dialect. For example:

```python
mbqc_pipeline = [
    (
        "default-pipeline",
        [
            "enforce-runtime-invariants-pipeline",
            "hlo-lowering-pipeline",
            "quantum-compilation-pipeline",
            "bufferization-pipeline",
        ],
    ),
    (
        "mbqc-pipeline",
        [
            "convert-mbqc-to-llvm",
        ],
    ),
    (
        "llvm-dialect-lowering-pipeline",
        [
            "llvm-dialect-lowering-pipeline",
        ],
    ),
]

@qjit(pipelines=mbqc_pipeline)
@qml.qnode(dev)
def workload():
    ...
```

No one is going to want to type that out each time, so it would be handy
to have this pipeline available in a common location.

**Description of the Change:** Adds a new function, `mbqc_pipeline()` to
a new `catalyst.ftqc` module. This function first makes a copy of the
default pipeline stages (from `catalyst.pipelines.default_pipeline()`,
recently added in #1941) and inserts the mbqc-to-llvm conversion pass
immediately before the quantum-to-llvm conversion pass in the
`"MLIRToLLVMDialect"` stage.

**Example:**

```python
import numpy as np
import pennylane as qml
from catalyst.ftqc import mbqc_pipeline

qml.capture.enable()

dev = qml.device("null.qubit", wires=1)


@qml.qjit(pipelines=mbqc_pipeline())
@qml.qnode(dev, mcm_method="tree-traversal")
def workload():
    m0 = qml.ftqc.measure_arbitrary_basis(wires=0, angle=np.pi / 4, plane="XY")
    qml.cond(m0, qml.X, qml.I)(0)
    return qml.expval(qml.Z(0))

workload()
```

**Benefits:** Less typing to specify the pass pipeline for MBQC
workloads (and therefore less error-prone!)

**Possible Drawbacks:** There is planned development work to more
broadly support user-definition of custom compilation pipelines in
Catalyst. Once this work is complete, it may be beneficial to revert
this change and use the new infrastructure.

**Note:** This PR was adapted from and supersedes #1758.

---------

Co-authored-by: Mehrdad Malek <39844030+mehrdad2m@users.noreply.github.com>