Fix qml.prod operator in AutoGraph decorator passthroughs (#1910)
Relies on https://github.com/PennyLaneAI/pennylane/pull/7889.

The above fixes the issue of AutoGraph transforming internal PennyLane
library code when it shouldn't. This PR provides an additional fix to
ensure AutoGraph doesn't fail when `qml.prod` operators are passed to
other decorators like `qml.adjoint`.

We have a fix in-place that transparently passes through decorators like
`qml.adjoint` without transforming the decorator, but still transforming
the function being wrapped by it. That code expects the first argument
to `qml.adjoint`-like functions to be either a PennyLane operation, or a
callable. However, `qml.prod` is a valid target that is an `Operator`
but not an `Operation`.
This PR changes the expected type to `Operator` (which is a parent class
of `Operation`).

NOTE: This PR does not enable support for AutoGraph to transform code
targeted by `qml.prod`, that is a separate issue.

closes #1824
[sc-93971]